% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-connectivity.R
\name{summarise-connectivity}
\alias{summarise-connectivity}
\alias{summarise_connectivity}
\alias{summarise_connectivity.default}
\title{Summarise connectivity metrics}
\usage{
summarise_connectivity(connectivity, ...)

\method{summarise_connectivity}{default}(
  connectivity,
  interpatch_distance,
  data_resolution,
  species,
  ...
)
}
\arguments{
\item{connectivity}{data.frame of class "patch_size_tbl", obtained via
\code{\link[=patch_sizes]{patch_sizes()}} from a \code{connectivity} object returned by
\code{\link[=habitat_connectivity]{habitat_connectivity()}}. Contains area measurements of connected patches.}

\item{...}{extra arguments to pass through for default method.}

\item{interpatch_distance}{Numeric. The distance (in meters) where habitat
patches are considered connected. E.g., if set to 500, patches 498m apart
are connected, those 501m apart are not connected. This is passed
internally to a spatial operation known as "buffering", where this
distance is used as a radius from the edge of the habitat zone. This means
the specified \code{interpatch_distance} is halved exactly. So an interpatch
distance of 500 will be converted to 250.}

\item{data_resolution}{Numeric. Data resolution in meters.}

\item{species}{Character. Name of species analysed.}
}
\value{
A tibble with connectivity metrics including number of patches,
probability of connectedness, effective mesh size, mean and total patch
areas.
}
\description{
Calculates a comprehensive set of habitat connectivity metrics including
effective mesh size, probability of connectedness, and patch statistics.
There are two methods, the default, and one that dispatches on objects of
class, \code{patch_size_tbl}, created by \code{\link[=patch_size_tbl]{patch_size_tbl()}}, which is used mostly
internally inside of \code{\link[=habitat_connectivity]{habitat_connectivity()}}. The default method requires
"connectivity" to be a vector of areas, and also requires scalar values
(values of length 1) of interpatch distance, resolution, and species. See
examples below.
}
\examples{
# dispatch on `patch_size_tbl` class - `lizard_areas_connected`
summarise_connectivity(
  connectivity = lizard_areas_connected
)

# default method
# manually passing area, interpatch distance, resolution, and species
summarise_connectivity(
  connectivity = lizard_areas_connected$area,
  interpatch_distance = 10,
  data_resolution = 10,
  species = "Blue-tongued Lizard"
)
}
